pyproject.toml
[project]
name = "qq-bot"
version = "0.1.0"
description = "QQ 聊天机器人，接入智谱 GLM 大模型"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "qq-botpy>=1.2.0,<2.0.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "loguru>=0.7.0,<1.0.0",
]

[project.scripts]
qq-bot = "qq_bot.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["qq_bot"]
